fix(frontend): sort admin config sessions with main before embedded
Extract shared useAdminSessions hook that applies sortSessionsByDate
to ensure deterministic ordering when sessions share the same
start_date (e.g., Session 2 before Session 2a). Both
GradeEligibilityConfig and SessionBudgetConfig now use this shared
hook instead of duplicate local useSessions with sort-by-date-only.
